requirements.txt
amqp==5.1.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
bleach==6.0.0
cachetools==5.3.1
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.1
comm==0.1.3
dask==2023.5.1
debugpy==1.6.7
decorator==5.1.1
Django==4.2.1
django-crontab==0.7.1
djangorestframework==3.14.0
executing==1.2.0
fsspec==2023.5.0
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.23.1
ipython==8.14.0
ipywidgets==8.0.6
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kombu==5.2.4
locket==1.0.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.2
parso==0.8.3
partd==1.4.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
prompt-toolkit==3.0.38
protobuf==4.23.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
swifter==1.3.4
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
widgetsnbextension==4.0.7
zipp==3.15.0
